Create RoombaGUIV1.py
Version 1 of our GUI. It uses Tiknter and features the components of our initial sketch of the GUI, including a camera toggle, compass, space for a camera feed, and joysticks to control the Roomba. However, some placeholders were used as many of these components will not be implemented for a while.