Create map.html
An extension to index.html. Messing with the point and click map feature but did not want to interfere with the working index.html file. 

For now, I have figured out how to have a point and click map typically seen in navigation, still need to incorporate data collection from OptiTrack and NatNet.